feat: allow outputting the workdir(s) for one or more Dockerfiles (#9)
* feat: allow outputting the workdir(s) for one or more Dockerfiles

This command is useful in inspection of Dockerfiles, and can be used to intelligently push/pull built files from a docker image. It also allows disabling of ranking, such that the the workdirs are output in order of appearance.

For instance, if a build tool expects to pull a config file from a container and has access to the respective Dockerfile, it can do something like:

```shell
WORKDIR=$(dockfmt workdir --no-rank --skip-header path/to/Dockerfile | tail -n1)
[[ -z "$WORKDIR" ]] && WORKDIR="/app"
docker cp "container-name:$WORKDIR/nginx.conf.template" nginx.conf || true
```

In this way, we can pull a file from a container at the best path possible - either a specified WORKDIR or a sane default.

* fix: correct lint errors

* chore: drop skip-header

This should be in a separate PR anyhow.